rename 'multi_skill' module to 'skill'
